updated rest client, adapted most screens to the new client

+++ app/src/main/java/de/andrestefanov/android/nearbuy/Preferences.kt
@@ -16,11 +16,11 @@ class Preferences {
 
         fun getToken(context: Context) = getPrefs(context).getString("token", null)
 
-        fun setUserId(context: Context, userId: Long) {
-            getPrefs(context).edit().putLong("userId", userId).apply()
+        fun setUserId(context: Context, userId: Int) {
+            getPrefs(context).edit().putInt("userId", userId).apply()
         }
 
-        fun getUserId(context: Context) = getPrefs(context).getLong("userId", -1)
+        fun getUserId(context: Context) = getPrefs(context).getInt("userId", -1)
 
     }
